Retrieve rowid for logged in user
Diplay messages in center

+++ src/auth.h
@@ -5,8 +5,8 @@
 #include <sqlite3.h>
 
 // Function declarations
-int signin (const char *username, const char *password, WINDOW *win);
+int64_t signin (const char *username, const char *password, WINDOW *win);
 int signup (const char *username, const char *password, WINDOW *win);
-int get_credentials_prompt (int action);
+int64_t get_credentials_prompt (int action);
 
 #endif // AUTH_H
